Rename colored.py to coloring.py
